Update dotnet.yml
I Found this super useful Report Generator that i saw other Groups use and asked them about it it basically takes our xml make a coverage report and puts it on a website  and automatically hides non happy paths. To USE this we need to fix the test job but if it works it will be really usefull thats why i added an extra branch.